requirements.txt
google-adk>=0.2.0
markdown
python-docx
fastapi>=0.104.0
uvicorn>=0.23.2
google-generativeai>=0.3.0
asyncio
python-multipart>=0.0.6
jinja2>=3.1.2
websockets>=11.0.3
firebase_admin
litellm
python-magic>=0.4.27
